+++ recipes/tket/conanfile.py
@@ -20,7 +20,7 @@ import shutil
 
 class TketConan(ConanFile):
     name = "tket"
-    version = "1.0.2"
+    version = "1.0.1"
     license = "CQC Proprietary"
     homepage = "https://github.com/CQCL/tket"
     url = "https://github.com/conan-io/conan-center-index"
@@ -36,7 +36,7 @@ class TketConan(ConanFile):
     exports_sources = ["../../tket/src/*", "!*/build/*"]
     requires = (
         # tk* libraries may come from remote:
-        # https://quantinuumsw.jfrog.io/artifactory/api/conan/tket1-libs
+        # https://tket.jfrog.io/artifactory/api/conan/tket-conan
         "boost/1.79.0",
         "symengine/0.9.0",
         "eigen/3.4.0",
@@ -45,7 +45,7 @@ class TketConan(ConanFile):
         "tkassert/0.1.1@tket/stable",
         "tkrng/0.1.2@tket/stable",
         "tktokenswap/0.1.1@tket/stable",
-        "tkwsm/0.1.1@tket/stable",
+        "tkwsm/0.2.0@tket/stable",
     )
 
     # List of components in a topological sort according to dependencies: