+++ tket/src/Gate/Gate.cpp
@@ -617,17 +617,78 @@ std::vector<Expr> Gate::get_params() const { return params_; }
 
 SymSet Gate::free_symbols() const { return expr_free_symbols(get_params()); }
 
+/**
+ * @brief The commutation colour of TK1(a,b,c)
+ *
+ * - commutes with everything if b == 0 && a == -c
+ * - commutes with Z if b == 0
+ * - commutes with X if a == c == 0
+ * - commutes with nothing otherwise
+ *
+ * @param params the TK1 angles
+ * @return std::optional<Pauli> commutation colour
+ */
+static std::optional<Pauli> tk1_commuting_basis(std::vector<Expr> params) {
+  const bool p0_is_zero = equiv_0(params[0], 2);
+  const bool p1_is_zero = equiv_0(params[1], 2);
+  const bool p2_is_zero = equiv_0(params[2], 2);
+  if (p1_is_zero && equiv_expr(params[0], -params[2], 2)) {
+    return Pauli::I;
+  } else if (p1_is_zero) {
+    return Pauli::Z;
+  } else if (p0_is_zero && p2_is_zero) {
+    return Pauli::X;
+  } else {
+    return std::nullopt;
+  }
+}
+
+/**
+ * @brief The commutation colour of TK2(a,b,c)
+ *
+ * - commutes with everything if a == b == c
+ * - commutes with X if b == c == 0
+ * - commutes with Y if a == c == 0
+ * - commutes with Z if a == b == 0
+ * - commutes with nothing otherwise
+ *
+ * @param params the TK2 angles
+ * @return std::optional<Pauli> commutation colour
+ */
+static std::optional<Pauli> tk2_commuting_basis(std::vector<Expr> params) {
+  const bool p0_is_zero = equiv_0(params[0], 2);
+  const bool p1_is_zero = equiv_0(params[1], 2);
+  const bool p2_is_zero = equiv_0(params[2], 2);
+  if (p0_is_zero && p1_is_zero && p2_is_zero) {
+    return Pauli::I;
+  } else if (p1_is_zero && p2_is_zero) {
+    return Pauli::X;
+  } else if (p0_is_zero && p2_is_zero) {
+    return Pauli::Y;
+  } else if (p0_is_zero && p1_is_zero) {
+    return Pauli::Z;
+  } else {
+    return std::nullopt;
+  }
+}
+
 std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
   unsigned n_q = n_qubits();
   if (i >= n_q) throw std::domain_error("Qubit index out of range");
   switch (get_type()) {
-    case OpType::XXPhase:
-    case OpType::XXPhase3: {
-      return Pauli::X;
-    }
-    case OpType::YYPhase: {
-      return Pauli::Y;
-    }
+    case OpType::X:
+    case OpType::V:
+    case OpType::Vdg:
+    case OpType::SX:
+    case OpType::SXdg:
+    case OpType::Rx:
+    case OpType::Y:
+    case OpType::Ry:
+    case OpType::noop:
+    case OpType::H:
+    case OpType::U3:
+    case OpType::U2:
+    case OpType::PhasedX:
     case OpType::Z:
     case OpType::S:
     case OpType::Sdg:
@@ -635,6 +696,16 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
     case OpType::Tdg:
     case OpType::Rz:
     case OpType::U1:
+    case OpType::TK1: {
+      return tk1_commuting_basis(get_tk1_angles());
+    }
+    case OpType::XXPhase:
+    case OpType::XXPhase3: {
+      return Pauli::X;
+    }
+    case OpType::YYPhase: {
+      return Pauli::Y;
+    }
     case OpType::CZ:
     case OpType::CRz:
     case OpType::CU1:
@@ -644,17 +715,11 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
     case OpType::CnZ: {
       return Pauli::Z;
     }
-    case OpType::noop: {
-      return Pauli::I;
-    }
-    case OpType::H:
-    case OpType::U3:
-    case OpType::U2:
-    case OpType::PhasedX:
-    case OpType::NPhasedX:
-    case OpType::TK1:
-    case OpType::TK2: {
+    case OpType::NPhasedX: {
       return std::nullopt;
+    }
+    case OpType::TK2: {
+      return tk2_commuting_basis(params_);
     }
     case OpType::CH:
     case OpType::CU3:
@@ -674,16 +739,10 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
         return Pauli::I;
       }
     }
-    case OpType::X:
-    case OpType::V:
-    case OpType::Vdg:
     case OpType::CV:
     case OpType::CVdg:
-    case OpType::SX:
-    case OpType::SXdg:
     case OpType::CSX:
     case OpType::CSXdg:
-    case OpType::Rx:
     case OpType::CRx:
     case OpType::CX:
     case OpType::CCX:
@@ -701,8 +760,6 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
         return std::nullopt;
       }
     }
-    case OpType::Y:
-    case OpType::Ry:
     case OpType::CY:
     case OpType::CRy:
     case OpType::CnRy:
@@ -721,11 +778,13 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
 
 bool Gate::commutes_with_basis(
     const std::optional<Pauli>& colour, unsigned i) const {
-  if (colour == Pauli::I) return true;
   const std::optional<Pauli> my_colour = commuting_basis(i);
-  if (!colour && !my_colour) return false;
-  if (my_colour == Pauli::I || my_colour == colour) return true;
-  return false;
+  if (!colour || !my_colour) {
+    // Nothing can commute with std::nullopt
+    return false;
+  } else {
+    return colour == Pauli::I || my_colour == Pauli::I || colour == my_colour;
+  }
 }
 
 op_signature_t Gate::get_signature() const {