+++ pytket/docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+1.5.2 (August 2022)
+-------------------
+
+Minor new features:
+
+* Prefer `ZZPhase` in ``DecomposeTK2`` if it results in the same fidelity but
+  fewer two-qubit gates.
+* New ``ToffoliBox`` for constructing circuits that implement permutations of
+  basis states.
+
 1.5.1 (August 2022)
 -------------------